Refactor: extract key objects for tables into their own files
