+++ app/assets/javascripts/utils/wordcount_utils.js
@@ -1,0 +1,3 @@
+export const bytesToWords = (bytes, bytesPerWord) => {
+  return Math.round(bytes / bytesPerWord);
+};
